Revert "Fix some compilation errors due to the switch to long character IDs"
This reverts commit 900d28ffbb2f2b2dc4344621089d1b05e3edbd0a.

+++ WaywardEconomy/src/main/java/net/wayward_realms/waywardeconomy/auction/AuctionManager.java
@@ -16,7 +16,7 @@ import java.util.Map;
 public class AuctionManager {
 
     private WaywardEconomy plugin;
-    private Map<Long, Auction> auctions = new HashMap<>();
+    private Map<Integer, Auction> auctions = new HashMap<>();
 
     public AuctionManager(WaywardEconomy plugin) {
         this.plugin = plugin;
